Feature/fs 114 add common mixin

+++ lib/filestack/mixins/filestack_common.rb
@@ -1,49 +1,78 @@
 require 'filestack/config'
+require 'filestack/utils/utils'
+require 'filestack/utils/multipart_upload_utils'
+require 'mimemagic'
 require 'typhoeus'
 
 # Module is mixin for common functionalities that all Filestack
 # objects can call.
 module FilestackCommon
-  # Send the delete api request to delete a filelink
+  include UploadUtils
+  # Get the contents of a Filestack handle
   #
   # @param [String]             file_handle   The Filelink handle
-  # @param [String]             apikey        Your Filestack API Key
   # @param [FilestackSecurity]  security      Optional Filestack security object
-  #                                           if security is enabled.
+  #                                           if security is enabled
   #
-  # @return [Typhoeus::Response]
-  def send_delete(file_handle, apikey, security = nil)
-    url = format('%s/file/%s?key=%s',
-                 FilestackConfig::API_URL,
-                 file_handle,
-                 apikey)
+  # @return [Bytes]
+  def send_get_content(url, parameters: nil)
+    UploadUtils.make_call(url, 'get', parameters: parameters)
+  end
 
-    url = security.sign_url(url) unless security.nil?
+  # Get the content of a filehandle
+  #
+  # @param [String]             filepath      Local path of file to be written
+  #
+  # @return [Int]                             Size of file written
+  def send_download(filepath)
+    content = send_get_content(url)
+    puts filepath
+    File.write(filepath, content.body)
+  end
 
-    Typhoeus.delete(url)
+  # Send the delete api request to delete a filelink
+  #
+  # @param [String]             url           The url of the Filehandle
+  #                                           to be deleted (includes security)
+  #
+  # @return [unirest::Response]
+  def send_delete(handle, apikey, security)
+    return 'Delete requires security' if security.nil?
+    signature = security.signature
+    policy = security.policy
+    base = "#{FilestackConfig::API_URL}/file"
+    UploadUtils.make_call(
+      "#{base}/#{handle}?key=#{apikey}&signature=#{signature}&policy=#{policy}",
+      'delete'
+    )
   end
 
   # Send the overwrite api request to update a filelink
   #
-  # @param [String]             filepath      filepath of file to upload
-  # @param [String]             file_handle   The Filelink handle
-  # @param [String]             apikey        Your Filestack API Key
-  # @param [FilestackSecurity]  security      Optional Filestack security object
-  #                                           if security is enabled.
+  # @param [String]             filepath      Filepath of file to upload
+  # @param [String]             handle        The Filelink handle
+  # @param [String]             apikey        Filestack API Key
+  # @param [FilestackSecurity]  security      Filestack security object
   #
   # @return [Typhoeus::Response]
-  def send_overwrite(filepath, file_handle, apikey, security = nil)
-    url = format('%s/file/%s?key=%s', FilestackConfig::API_URL,
-                 file_handle, apikey)
-    url = security.sign_url(url) unless security.nil?
+  def send_overwrite(filepath, handle, apikey, security)
+    return 'Overwrite requires security' if security.nil?
 
     file = File.open(filepath, 'r')
-    contents = file.read
-    headers = { 'Content-Type' => 'application/octet-stream' }
+    mimetype = MimeMagic.by_magic(file)
+    content = file.read
 
-    request = Typhoeus::Request.new(url, headers: headers,
-                                         method: :post,
-                                         body: contents)
-    request.run
+    signature = security.signature
+    policy = security.policy
+
+    headers = { 'Content-Type' => mimetype }
+    base = "#{FilestackConfig::API_URL}/file"
+
+    UploadUtils.make_call(
+      "#{base}/#{handle}?key=#{apikey}&signature=#{signature}&policy=#{policy}",
+      'put',
+      headers: headers,
+      parameters: content
+    )
   end
 end
